refactor: Config 파일 분리
